refactor: replace inline styles with css module in the not-found component
